admin cardrule some feature&template feature

+++ src/main/java/com/by/repository/RuleRepository.java
@@ -1,9 +1,8 @@
 package com.by.repository;
 
-import com.by.model.Rule;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Pageable;
 import org.springframework.data.repository.PagingAndSortingRepository;
 
-public interface RuleRepository extends PagingAndSortingRepository<Rule, Long> {
+import com.by.model.Rule;
+
+public interface RuleRepository extends PagingAndSortingRepository<Rule, Integer> {
 }
